Rename to RequestForm/FormField
The name "Contract" was used to avoid conflicts with the legacy form
classes, but since those aren't in the public release anyways, there's
not as much need to avoid it. However, these are primarily intended for
request processing, with the additional capability of being used for
templating, so the name RequestForm made more sense than simply "Form".